requirements.txt
redis==5.2.0
temporalio==1.8.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.10
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
tenacity==9.0.0